+++ src/test/java/eu/verdelhan/ta4j/indicator/oscillator/AwesomeOscillatorTest.java
@@ -32,33 +32,33 @@ public class AwesomeOscillatorTest {
 	public void testCalculateWithSma2AndSma3() throws Exception {
 		AwesomeOscillator awesome = new AwesomeOscillator(new AverageHighLow(series), 2, 3);
 
-		assertThat((double) awesome.getValue(0)).isEqualTo(0d);
-		assertThat((double) awesome.getValue(1)).isEqualTo(0d);
+		assertThat(awesome.getValue(0)).isEqualTo(0d);
+		assertThat(awesome.getValue(1)).isEqualTo(0d);
 		assertThat(awesome.getValue(2)).isEqualTo(0.1666666d);
 		assertThat(awesome.getValue(3)).isEqualTo(1d);
-		assertThat((double) awesome.getValue(4)).isEqualTo(-3d);
+		assertThat(awesome.getValue(4)).isEqualTo(-3d);
 	}
 
 	@Test
 	public void testWithSma1AndSma2() throws Exception {
 		AwesomeOscillator awesome = new AwesomeOscillator(new AverageHighLow(series), 1, 2);
 
-		assertThat((double) awesome.getValue(0)).isEqualTo(0d);
-		assertThat((double) awesome.getValue(1)).isEqualTo(-1.5d);
-		assertThat((double) awesome.getValue(2)).isEqualTo(3.5d);
-		assertThat((double) awesome.getValue(3)).isEqualTo(-4d);
-		assertThat((double) awesome.getValue(4)).isEqualTo(-1d);
+		assertThat(awesome.getValue(0)).isEqualTo(0d);
+		assertThat(awesome.getValue(1)).isEqualTo(-1.5d);
+		assertThat(awesome.getValue(2)).isEqualTo(3.5d);
+		assertThat(awesome.getValue(3)).isEqualTo(-4d);
+		assertThat(awesome.getValue(4)).isEqualTo(-1d);
 	}
 
 	@Test
 	public void testWithSmaDefault() throws Exception {
 		AwesomeOscillator awesome = new AwesomeOscillator(new AverageHighLow(series));
 
-		assertThat((double) awesome.getValue(0)).isEqualTo(0d);
-		assertThat((double) awesome.getValue(1)).isEqualTo(0d);
-		assertThat((double) awesome.getValue(2)).isEqualTo(0d);
-		assertThat((double) awesome.getValue(3)).isEqualTo(0d);
-		assertThat((double) awesome.getValue(4)).isEqualTo(0d);
+		assertThat(awesome.getValue(0)).isEqualTo(0d);
+		assertThat(awesome.getValue(1)).isEqualTo(0d);
+		assertThat(awesome.getValue(2)).isEqualTo(0d);
+		assertThat(awesome.getValue(3)).isEqualTo(0d);
+		assertThat(awesome.getValue(4)).isEqualTo(0d);
 	}
 
 }
