+++ src/test/java/eu/verdelhan/ta4j/indicator/tracker/WMATest.java
@@ -15,12 +15,12 @@ public class WMATest {
 		Indicator<BigDecimal> close = new ClosePrice(series);
 		Indicator<Double> wmaIndicator = new WMA(close, 3);
 		
-		assertThat((double) wmaIndicator.getValue(0)).isEqualTo(1d);
-		assertThat((double) wmaIndicator.getValue(1)).isEqualTo(5d/3);
-		assertThat((double) wmaIndicator.getValue(2)).isEqualTo(14d/6);
-		assertThat((double) wmaIndicator.getValue(3)).isEqualTo(20d/6);
-		assertThat((double) wmaIndicator.getValue(4)).isEqualTo(26d/6);
-		assertThat((double) wmaIndicator.getValue(5)).isEqualTo(32d/6);
+		assertThat(wmaIndicator.getValue(0)).isEqualTo(1d);
+		assertThat(wmaIndicator.getValue(1)).isEqualTo(5d/3);
+		assertThat(wmaIndicator.getValue(2)).isEqualTo(14d/6);
+		assertThat(wmaIndicator.getValue(3)).isEqualTo(20d/6);
+		assertThat(wmaIndicator.getValue(4)).isEqualTo(26d/6);
+		assertThat(wmaIndicator.getValue(5)).isEqualTo(32d/6);
 	}
 	
 	@Test
@@ -30,7 +30,7 @@ public class WMATest {
 		Indicator<BigDecimal> close = new ClosePrice(series);
 		Indicator<Double> wmaIndicator = new WMA(close, 3);
 		
-		assertThat((double) wmaIndicator.getValue(5)).isEqualTo(32d/6);
+		assertThat(wmaIndicator.getValue(5)).isEqualTo(32d/6);
 	}
 	
 	@Test
@@ -40,12 +40,12 @@ public class WMATest {
 		Indicator<BigDecimal> close = new ClosePrice(series);
 		Indicator<Double> wmaIndicator = new WMA(close, 55);
 		
-		assertThat((double) wmaIndicator.getValue(0)).isEqualTo(1d);
-		assertThat((double) wmaIndicator.getValue(1)).isEqualTo(5d/3);
-		assertThat((double) wmaIndicator.getValue(2)).isEqualTo(14d/6);
-		assertThat((double) wmaIndicator.getValue(3)).isEqualTo(30d/10);
-		assertThat((double) wmaIndicator.getValue(4)).isEqualTo(55d/15);
-		assertThat((double) wmaIndicator.getValue(5)).isEqualTo(91d/21);
+		assertThat(wmaIndicator.getValue(0)).isEqualTo(1d);
+		assertThat(wmaIndicator.getValue(1)).isEqualTo(5d/3);
+		assertThat(wmaIndicator.getValue(2)).isEqualTo(14d/6);
+		assertThat(wmaIndicator.getValue(3)).isEqualTo(30d/10);
+		assertThat(wmaIndicator.getValue(4)).isEqualTo(55d/15);
+		assertThat(wmaIndicator.getValue(5)).isEqualTo(91d/21);
 	}
 
 }
